+++ internal/provider/data_source_ise_internal_user_test.go
@@ -45,7 +45,6 @@ func TestAccDataSourceIseInternalUser(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.ise_internal_user.test", "first_name", "John"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.ise_internal_user.test", "last_name", "Doe"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.ise_internal_user.test", "password_id_store", "Internal Users"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.ise_internal_user.test", "expiry_date_enabled", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.ise_internal_user.test", "description", "My first Terraform user"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
@@ -82,7 +81,6 @@ func testAccDataSourceIseInternalUserConfig() string {
 	config += `	first_name = "John"` + "\n"
 	config += `	last_name = "Doe"` + "\n"
 	config += `	password_id_store = "Internal Users"` + "\n"
-	config += `	expiry_date_enabled = false` + "\n"
 	config += `	description = "My first Terraform user"` + "\n"
 	config += `}` + "\n"