+++ internal/provider/model_ise_internal_user.go
@@ -46,8 +46,6 @@ type InternalUser struct {
 	IdentityGroups       types.String `tfsdk:"identity_groups"`
 	CustomAttributes     types.String `tfsdk:"custom_attributes"`
 	PasswordIdStore      types.String `tfsdk:"password_id_store"`
-	ExpiryDateEnabled    types.Bool   `tfsdk:"expiry_date_enabled"`
-	ExpiryDate           types.String `tfsdk:"expiry_date"`
 	Description          types.String `tfsdk:"description"`
 }
 
@@ -101,12 +99,6 @@ func (data InternalUser) toBody(ctx context.Context, state InternalUser) string 
 	}
 	if !data.PasswordIdStore.IsNull() {
 		body, _ = sjson.Set(body, "InternalUser.passwordIDStore", data.PasswordIdStore.ValueString())
-	}
-	if !data.ExpiryDateEnabled.IsNull() {
-		body, _ = sjson.Set(body, "InternalUser.expiryDateEnabled", data.ExpiryDateEnabled.ValueBool())
-	}
-	if !data.ExpiryDate.IsNull() {
-		body, _ = sjson.Set(body, "InternalUser.expiryDate", data.ExpiryDate.ValueString())
 	}
 	if !data.Description.IsNull() {
 		body, _ = sjson.Set(body, "InternalUser.description", data.Description.ValueString())
@@ -173,16 +165,6 @@ func (data *InternalUser) fromBody(ctx context.Context, res gjson.Result) {
 	} else {
 		data.PasswordIdStore = types.StringValue("Internal Users")
 	}
-	if value := res.Get("InternalUser.expiryDateEnabled"); value.Exists() {
-		data.ExpiryDateEnabled = types.BoolValue(value.Bool())
-	} else {
-		data.ExpiryDateEnabled = types.BoolValue(false)
-	}
-	if value := res.Get("InternalUser.expiryDate"); value.Exists() {
-		data.ExpiryDate = types.StringValue(value.String())
-	} else {
-		data.ExpiryDate = types.StringNull()
-	}
 	if value := res.Get("InternalUser.description"); value.Exists() {
 		data.Description = types.StringValue(value.String())
 	} else {
@@ -248,16 +230,6 @@ func (data *InternalUser) updateFromBody(ctx context.Context, res gjson.Result) 
 		data.PasswordIdStore = types.StringValue(value.String())
 	} else if data.PasswordIdStore.ValueString() != "Internal Users" {
 		data.PasswordIdStore = types.StringNull()
-	}
-	if value := res.Get("InternalUser.expiryDateEnabled"); value.Exists() && !data.ExpiryDateEnabled.IsNull() {
-		data.ExpiryDateEnabled = types.BoolValue(value.Bool())
-	} else if data.ExpiryDateEnabled.ValueBool() != false {
-		data.ExpiryDateEnabled = types.BoolNull()
-	}
-	if value := res.Get("InternalUser.expiryDate"); value.Exists() && !data.ExpiryDate.IsNull() {
-		data.ExpiryDate = types.StringValue(value.String())
-	} else {
-		data.ExpiryDate = types.StringNull()
 	}
 	if value := res.Get("InternalUser.description"); value.Exists() && !data.Description.IsNull() {
 		data.Description = types.StringValue(value.String())