+++ docs/guides/changelog.md
@@ -12,7 +12,7 @@ description: |-
 - Fix issue with reading nested conditions using `ise_network_access_condition` resource
 - Add support for nested conditions to `ise_network_access_policy_set` resource and data source
 - Add `agentless_posture` attribute to `ise_authorization_profile` resource and data source
-- Add `identity_groups`, `custom_attributes`, `expiry_date_enabled` and `expiry_date` attributes to `ise_internal_user` resource and data source
+- Add `identity_groups` and `custom_attributes` attributes to `ise_internal_user` resource and data source
 - Add support for nested conditions to `ise_network_access_authentication_rule` resource and data source
 - Add support for nested conditions to `ise_network_access_authorization_rule` resource and data source
 - Add `ise_trustsec_egress_matrix_cell` resource and data source